+++ src/platform_impl/macos/view.rs
@@ -238,6 +238,10 @@ lazy_static! {
             scroll_wheel as extern "C" fn(&Object, Sel, id),
         );
         decl.add_method(
+            sel!(magnifyWithEvent:),
+            magnify_with_event as extern "C" fn(&Object, Sel, id),
+        );
+        decl.add_method(
             sel!(pressureChangeWithEvent:),
             pressure_change_with_event as extern "C" fn(&Object, Sel, id),
         );
@@ -1044,6 +1048,37 @@ extern "C" fn scroll_wheel(this: &Object, _sel: Sel, event: id) {
         AppState::queue_event(EventWrapper::StaticEvent(window_event));
     }
     trace!("Completed `scrollWheel`");
+}
+
+extern "C" fn magnify_with_event(this: &Object, _sel: Sel, event: id) {
+    trace!("Triggered `magnifyWithEvent`");
+
+    mouse_motion(this, event);
+
+    unsafe {
+        let state_ptr: *mut c_void = *this.get_ivar("winitState");
+        let state = &mut *(state_ptr as *mut ViewState);
+
+        let delta = event.magnification();
+        let phase = match event.phase() {
+            NSEventPhase::NSEventPhaseBegan => TouchPhase::Started,
+            NSEventPhase::NSEventPhaseChanged => TouchPhase::Moved,
+            NSEventPhase::NSEventPhaseEnded => TouchPhase::Ended,
+            _ => return
+        };
+
+        let window_event = Event::WindowEvent {
+            window_id: WindowId(get_window_id(state.ns_window)),
+            event: WindowEvent::TouchpadMagnify {
+                device_id: DEVICE_ID,
+                delta,
+                phase,
+            },
+        };
+
+        AppState::queue_event(EventWrapper::StaticEvent(window_event));
+    }
+    trace!("Completed `magnifyWithEvent`");
 }
 
 extern "C" fn pressure_change_with_event(this: &Object, _sel: Sel, event: id) {